# openapi_reveal.yaml
openapi: 3.1.0
info:
  title: Kaveri Stays API
  version: "1.0.0"
  summary: Reveal specification for the Kaveri Stays API assignment.
  description: |
    # Read this before you read the paths

    This is ONE reasonable design. It is not the only one. Several of the choices
    below are genuinely arguable and the Interlude expects you to argue with them.
    A few are not arguable, and those are marked.

    ## Conventions this specification commits to

    **Dates.** Every date is ISO 8601 `YYYY-MM-DD`. A stay is the half-open interval
    `[check_in, check_out)`. The check-out date is NOT part of the stay. This is what
    makes same-day turnover work (rule 5), and it is NOT arguable.

    **Money.** Every amount is a decimal string, never a JSON number. `"4500.00"`,
    not `4500.0`. Floats lose money.

    **Empty collections.** A list endpoint with no matching rows returns 200 and an
    empty `items` array. It never returns 404. 404 means the container does not exist,
    not that it is empty.

    **401 versus 403.** 401 means the server does not know who you are: no token, an
    expired token, a forged token. 403 means it knows exactly who you are and you may
    not do this. This distinction is NOT arguable.

    **Request bodies.** Every request schema sets `additionalProperties: false`. A
    field you did not declare is rejected rather than ignored. This is what defeats
    attacks 8.2 and 8.8 structurally rather than by remembering to check.

...

...

...

...

...

  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Exchange credentials for a token pair.
      description: |
        The failure response is identical for an unknown email and a wrong password.
        Anything else answers attack 8.14 for the attacker.
      security: []
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /me:
    get:
      tags: [auth]
      operationId: getMe
      summary: The caller's own record.
      x-roles: [guest, staff, manager, owner]
      responses:
        '200':
          description: The caller.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ----------------------------------------------------------- properties ---
...
  /bookings/{booking_id}/review:
    post:
      tags: [reviews]
      operationId: createReview
      summary: Review a completed stay.
      description: |
        Task 5.8. Rule 9: only after checkout, only once, rating 1 to 5.

        Three distinct failures with three distinct codes. Reviewing someone else's
        stay is 404, consistent with `GET /bookings/{id}`. Reviewing your own stay
        before you have checked out is 403, because you personally are not yet
        permitted to do this. A second review on the same booking is 409, because it
        collides with a row that already exists. A rating of 9 is 422.
      x-roles: [guest]
      parameters:
        - $ref: '#/components/parameters/BookingIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewCreate' }
      responses:
        '201':
          description: Review recorded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: The stay is not checked out yet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: This booking already has a review.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /properties/{property_id}/reviews:
    get:
      tags: [reviews]
      operationId: listPropertyReviews
      summary: Reviews for a property.
      description: Guest names are shortened. Reviews never expose a guest email.
      security: []
      x-roles: [public]
      parameters:
        - $ref: '#/components/parameters/PropertyIdPath'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Reviews.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewPage' }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------------- reports ---
...
  /guests:
    get:
      tags: [guests]
      operationId: listGuests
      summary: Guest records.
      description: |
        Staff and above only. There is deliberately no endpoint through which a guest
        can list, search or count other guests. Task 4.8.
      x-roles: [staff, manager, owner]
      parameters:
        - name: email
          in: query
          description: Exact match, case-insensitive. Not a prefix search.
          schema: { type: string, format: email }
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Guests.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /guests/{guest_id}:
    get:
      tags: [guests]
      operationId: getGuest
      summary: One guest record.
      x-roles: [staff, manager, owner]
      parameters:
        - name: guest_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: The guest.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Guest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

# ========================================================== components ======
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token in `Authorization: Bearer <token>`. Short lived. Obtain a new one
        from `/auth/refresh` rather than by logging in again.
  parameters:
    PropertyIdPath:
      name: property_id
      in: path
      required: true
      schema: { type: integer }
    BookingIdPath:
      name: booking_id
      in: path
      required: true
      schema: { type: integer }
    LimitParam:
      name: limit
      in: query
      description: Page size. Capped server-side; an unbounded list endpoint is a defect.
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    Unauthorized:
      description: No token, an expired token, or a token that does not verify.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: unauthenticated, message: Authentication required. }
    Forbidden:
      description: Authenticated, and not permitted to do this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: forbidden, message: You do not have access to this resource. }
    NotFound:
      description: No such resource, or none you are permitted to know about.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: not_found, message: Not found. }
...
    Unprocessable:
      description: |
        The request parsed and a field is unacceptable. Includes guest counts above
        the room type maximum, check-out on or before check-in, ratings outside 1 to 5,
        and any field not declared in the request schema.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: validation_failed
              message: Request validation failed.
              detail:
                - field: guests
                  reason: Exceeds maximum occupancy of 2 for room type Standard.
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
      headers:
        Retry-After:
          description: Seconds until the caller may retry.
          schema: { type: integer }
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      description: |
        Every non-2xx response in this API uses this shape. One handler produces it.
        If you are constructing this dict in more than one module, consolidate.
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable, snake_case. Clients branch on this.
              examples: [room_unavailable, illegal_transition, validation_failed]
            message:
              type: string
              description: Human-readable. Safe to show a guest. Never a raw Postgres message.
            detail:
              description: Optional structured context. Never leaks table or column names.
            request_id:
              type: string
              description: Correlates this response with your logs.
    PageMeta:
      type: object
      additionalProperties: false
      required: [limit, offset, total]
      properties:
        limit:  { type: integer }
        offset: { type: integer }
        total:  { type: integer, description: Total matching rows, ignoring pagination. }

    # ---- enums ----
...
    Role:
      type: string
      enum: [guest, staff, manager, owner]

    # ---- auth ----
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string, writeOnly: true }
...
    TokenPair:
      type: object
      additionalProperties: false
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:  { type: string }
        refresh_token: { type: string }
        token_type:    { type: string, const: bearer }
        expires_in:    { type: integer, description: Access token lifetime in seconds. }
    Me:
      type: object
      additionalProperties: false
      required: [id, email, full_name, role]
      description: No password hash appears in this schema or any other. Task 4.6.
      properties:
        id:          { type: integer }
        email:       { type: string, format: email }
        full_name:   { type: string }
        role:        { $ref: '#/components/schemas/Role' }
        property_id:
          type: [integer, "null"]
          description: The property this staff member or manager belongs to. Null for guest and owner.

    # ---- properties and rooms ----
...
    ReviewCreate:
      type: object
      additionalProperties: false
      required: [rating]
      properties:
        rating:  { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, maxLength: 2000 }
    Review:
      type: object
      additionalProperties: false
      required: [id, booking_id, rating, created_at]
      properties:
        id:          { type: integer }
        booking_id:  { type: integer }
        rating:      { type: integer, minimum: 1, maximum: 5 }
        comment:     { type: string }
        guest_name:  { type: string, description: Shortened. Never the full record, never the email. }
        created_at:  { type: string, format: date-time }
    ReviewPage:
      type: object
      additionalProperties: false
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Review' }
        meta: { $ref: '#/components/schemas/PageMeta' }

    # ---- reports ----
...
    Guest:
      type: object
      additionalProperties: false
      required: [id, email, full_name]
      properties:
        id:         { type: integer }
        email:      { type: string, format: email }
        full_name:  { type: string }
        phone:      { type: string }
        stay_count: { type: integer }
    GuestPage:
      type: object
      additionalProperties: false
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Guest' }
        meta: { $ref: '#/components/schemas/PageMeta' }